refactor: took the bottom tabs page out
